+++ src/syrupy/__init__.py
@@ -1,4 +1,6 @@
+import argparse
 import glob
+from gettext import gettext
 from typing import (
     Any,
     List,
@@ -8,6 +10,7 @@ from typing import (
 import pytest
 
 from .assertion import SnapshotAssertion
+from .exceptions import FailedToLoadModuleMember
 from .extensions import DEFAULT_EXTENSION
 from .location import TestLocation
 from .session import SnapshotSession
@@ -16,6 +19,14 @@ from .terminal import (
     reset,
     snapshot_color,
 )
+from .utils import import_module_member
+
+
+def __default_extension_option(value: str) -> Any:
+    try:
+        return import_module_member(value)
+    except FailedToLoadModuleMember as e:
+        raise argparse.ArgumentTypeError(e)
 
 
 def pytest_addoption(parser: Any) -> None:
@@ -37,6 +48,13 @@ def pytest_addoption(parser: Any) -> None:
         default=False,
         dest="warn_unused_snapshots",
         help="Do not fail on unused snapshots",
+    )
+    group.addoption(
+        "--snapshot-default-extension",
+        type=__default_extension_option,
+        default=DEFAULT_EXTENSION,
+        dest="default_extension",
+        help="Specify the default snapshot extension",
     )
 
 
@@ -62,6 +80,10 @@ def __is_testpath(arg: str) -> bool:
     return not arg.startswith("-") and bool(glob.glob(arg.split("::")[0]))
 
 
+def __is_testnode(arg: str) -> bool:
+    return __is_testpath(arg) and "::" in arg
+
+
 def __is_testmodule(arg: str) -> bool:
     return arg == "--pyargs"
 
@@ -72,7 +94,7 @@ def pytest_sessionstart(session: Any) -> None:
     https://docs.pytest.org/en/latest/reference.html#_pytest.hookspec.pytest_sessionstart
     """
     config = session.config
-    session._syrupy = SnapshotSession(
+    config._syrupy = SnapshotSession(
         warn_unused_snapshots=config.option.warn_unused_snapshots,
         update_snapshots=config.option.update_snapshots,
         base_dir=config.rootdir,
@@ -80,8 +102,11 @@ def pytest_sessionstart(session: Any) -> None:
             __is_testpath(arg) or __is_testmodule(arg)
             for arg in config.invocation_params.args
         ),
+        is_providing_nodes=any(
+            __is_testnode(arg) for arg in config.invocation_params.args
+        ),
     )
-    session._syrupy.start()
+    config._syrupy.start()
 
 
 def pytest_collection_modifyitems(session: Any, config: Any, items: List[Any]) -> None:
@@ -89,7 +114,7 @@ def pytest_collection_modifyitems(session: Any, config: Any, items: List[Any]) -
     After tests are collected and before any modification is performed.
     https://docs.pytest.org/en/latest/reference.html#_pytest.hookspec.pytest_collection_modifyitems
     """
-    session._syrupy._all_items.update(items)
+    config._syrupy._all_items.update(items)
 
 
 def pytest_collection_finish(session: Any) -> None:
@@ -97,26 +122,34 @@ def pytest_collection_finish(session: Any) -> None:
     After collection has been performed and modified.
     https://docs.pytest.org/en/latest/reference.html#_pytest.hookspec.pytest_collection_finish
     """
-    session._syrupy._ran_items.update(session.items)
+    session.config._syrupy._ran_items.update(session.items)
 
 
 def pytest_sessionfinish(session: Any, exitstatus: int) -> None:
     """
-    Add syrupy report to pytest after whole test run finished, before exiting.
+    Finish session run and set exit status.
     https://docs.pytest.org/en/latest/reference.html#_pytest.hookspec.pytest_sessionfinish
     """
-    reporter = session.config.pluginmanager.get_plugin("terminalreporter")
-    syrupy_exitstatus = session._syrupy.finish()
-    for line in session._syrupy.report.lines:
-        reporter.write_line(line)
-    session.exitstatus |= syrupy_exitstatus
+    session.exitstatus |= exitstatus | session.config._syrupy.finish()
+
+
+def pytest_terminal_summary(
+    terminalreporter: Any, exitstatus: int, config: Any
+) -> None:
+    """
+    Add syrupy report to pytest.
+    https://docs.pytest.org/en/latest/reference.html#_pytest.hookspec.pytest_terminal_summary
+    """
+    terminalreporter.write_sep("-", gettext("snapshot report summary"))
+    for line in terminalreporter.config._syrupy.report.lines:
+        terminalreporter.write_line(line)
 
 
 @pytest.fixture
 def snapshot(request: Any) -> "SnapshotAssertion":
     return SnapshotAssertion(
         update_snapshots=request.config.option.update_snapshots,
-        extension_class=DEFAULT_EXTENSION,
+        extension_class=request.config.option.default_extension,
         test_location=TestLocation(request.node),
-        session=request.session._syrupy,
+        session=request.session.config._syrupy,
     )