+++ src/syrupy/exceptions.py
@@ -1,2 +1,6 @@
 class SnapshotDoesNotExist(Exception):
     """Snapshot does not exist"""
+
+
+class FailedToLoadModuleMember(Exception):
+    """Failed to load specific member in a module"""