+++ src/syrupy/extensions/base.py
@@ -1,4 +1,3 @@
-import os
 import warnings
 from abc import (
     ABC,
@@ -7,6 +6,7 @@ from abc import (
 from collections import deque
 from difflib import ndiff
 from gettext import gettext
+from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Deque,
@@ -68,7 +68,7 @@ class SnapshotFossilizer(ABC):
     def get_location(self, *, index: int) -> str:
         """Returns full location where snapshot data is stored."""
         basename = self._get_file_basename(index=index)
-        return os.path.join(self._dirname, f"{basename}.{self._file_extension}")
+        return str(Path(self._dirname).joinpath(f"{basename}.{self._file_extension}"))
 
     def is_snapshot_location(self, *, location: str) -> bool:
         """Checks if supplied location is valid for this snapshot extension"""
@@ -188,8 +188,8 @@ class SnapshotFossilizer(ABC):
 
     @property
     def _dirname(self) -> str:
-        test_dirname = os.path.dirname(self.test_location.filepath)
-        return os.path.join(test_dirname, SNAPSHOT_DIRNAME)
+        test_dir = Path(self.test_location.filepath).parent
+        return str(test_dir.joinpath(SNAPSHOT_DIRNAME))
 
     @property
     @abstractmethod
@@ -205,7 +205,7 @@ class SnapshotFossilizer(ABC):
         Ensures the folder path for the snapshot file exists.
         """
         try:
-            os.makedirs(os.path.dirname(self.get_location(index=index)))
+            Path(self.get_location(index=index)).parent.mkdir(parents=True)
         except FileExistsError:
             pass