+++ tests/test_integration_default.py
@@ -1,12 +1,19 @@
-import os
+from pathlib import Path
 
 import pytest
 
 from .utils import clean_output
 
 
+def get_result_snapshot_summary(result: object) -> str:
+    result_stdout = clean_output(result.stdout.str())
+    start_idx = result_stdout.find("\n", result_stdout.find("---- snapshot report"))
+    end_idx = result_stdout.find("====", start_idx)
+    return result_stdout[start_idx:end_idx].strip()
+
+
 @pytest.fixture
-def collection(testdir):
+def collection(testdir, snapshot):
     tests = {
         "test_collected": (
             """
@@ -19,20 +26,19 @@ def collection(testdir):
         ),
         "test_not_collected": (
             """
-            def test_not_collected(snapshot):
+            def test_collected1(snapshot):
                 assert snapshot == "hello"
             """
         ),
     }
     testdir.makepyfile(**tests)
     result = testdir.runpytest("-v", "--snapshot-update")
-    result_stdout = clean_output(result.stdout.str())
-    assert "4 snapshots generated" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     testdir.makefile(".ambr", **{"__snapshots__/other_snapfile": ""})
     return testdir
 
 
-def test_unused_snapshots_ignored_if_not_targeted_using_dash_m(collection):
+def test_unused_snapshots_ignored_if_not_targeted_using_dash_m(collection, snapshot):
     updated_tests = {
         "test_collected": (
             """
@@ -46,16 +52,13 @@ def test_unused_snapshots_ignored_if_not_targeted_using_dash_m(collection):
     }
     collection.makepyfile(**updated_tests)
     result = collection.runpytest("-v", "--snapshot-update", "-m", "parametrize")
-    result_stdout = clean_output(result.stdout.str())
-    assert "1 snapshot passed" in result_stdout
-    assert "1 snapshot updated" in result_stdout
-    assert "1 unused snapshot deleted" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     snapshot_path = [collection.tmpdir, "__snapshots__"]
-    assert os.path.isfile(os.path.join(*snapshot_path, "test_not_collected.ambr"))
-    assert os.path.isfile(os.path.join(*snapshot_path, "other_snapfile.ambr"))
+    assert Path(*snapshot_path).joinpath("test_not_collected.ambr").exists()
+    assert Path(*snapshot_path).joinpath("other_snapfile.ambr").exists()
 
 
-def test_unused_snapshots_ignored_if_not_targeted_using_dash_k(collection):
+def test_unused_snapshots_ignored_if_not_targeted_using_dash_k(collection, snapshot):
     updated_tests = {
         "test_collected": (
             """
@@ -68,17 +71,16 @@ def test_unused_snapshots_ignored_if_not_targeted_using_dash_k(collection):
         ),
     }
     collection.makepyfile(**updated_tests)
-    result = collection.runpytest("-v", "--snapshot-update", "-k", "test_collected")
-    result_stdout = clean_output(result.stdout.str())
-    assert "1 snapshot passed" in result_stdout
-    assert "1 snapshot updated" in result_stdout
-    assert "1 unused snapshot deleted" in result_stdout
+    result = collection.runpytest("-v", "--snapshot-update", "-k", "test_collected[")
+    assert get_result_snapshot_summary(result) == snapshot
     snapshot_path = [collection.tmpdir, "__snapshots__"]
-    assert os.path.isfile(os.path.join(*snapshot_path, "test_not_collected.ambr"))
-    assert os.path.isfile(os.path.join(*snapshot_path, "other_snapfile.ambr"))
+    assert Path(*snapshot_path).joinpath("test_not_collected.ambr").exists()
+    assert Path(*snapshot_path).joinpath("other_snapfile.ambr").exists()
 
 
-def test_unused_parameterized_ignored_if_not_targeted_using_dash_k(collection):
+def test_unused_parameterized_ignored_if_not_targeted_using_dash_k(
+    collection, snapshot
+):
     updated_tests = {
         "test_collected": (
             """
@@ -91,14 +93,11 @@ def test_unused_parameterized_ignored_if_not_targeted_using_dash_k(collection):
         ),
     }
     collection.makepyfile(**updated_tests)
-    result = collection.runpytest("-v", "--snapshot-update", "-k", "test_collected")
-    result_stdout = clean_output(result.stdout.str())
-    assert "2 snapshots passed" in result_stdout
-    assert "snapshot updated" not in result_stdout
-    assert "1 unused snapshot deleted" in result_stdout
+    result = collection.runpytest("-v", "--snapshot-update", "-k", "test_collected[")
+    assert get_result_snapshot_summary(result) == snapshot
     snapshot_path = [collection.tmpdir, "__snapshots__"]
-    assert os.path.isfile(os.path.join(*snapshot_path, "test_not_collected.ambr"))
-    assert os.path.isfile(os.path.join(*snapshot_path, "other_snapfile.ambr"))
+    assert Path(*snapshot_path).joinpath("test_not_collected.ambr").exists()
+    assert Path(*snapshot_path).joinpath("other_snapfile.ambr").exists()
 
 
 @pytest.fixture
@@ -209,10 +208,10 @@ def testcases_updated(testcases):
     return {**testcases, **updated_testcases}
 
 
-def test_missing_snapshots(testdir, testcases):
+def test_missing_snapshots(testdir, testcases, snapshot):
     testdir.makepyfile(test_file=testcases["used"])
     result = testdir.runpytest("-v")
-    assert "Snapshot does not exist" in clean_output(result.stdout.str())
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 1
 
 
@@ -220,7 +219,7 @@ def test_missing_snapshots(testdir, testcases):
 def stubs(testdir, testcases):
     pyfile_content = "\n\n".join(testcases.values())
     testdir.makepyfile(test_file=pyfile_content)
-    filepath = os.path.join(testdir.tmpdir, "__snapshots__", "test_file.ambr")
+    filepath = Path(testdir.tmpdir, "__snapshots__", "test_file.ambr")
     return testdir.runpytest("-v", "--snapshot-update"), testdir, testcases, filepath
 
 
@@ -230,11 +229,9 @@ def test_injected_fixture(stubs):
     assert result.ret == 0
 
 
-def test_generated_snapshots(stubs):
+def test_generated_snapshots(stubs, snapshot):
     result = stubs[0]
-    result_stdout = clean_output(result.stdout.str())
-    assert "7 snapshots generated" in result_stdout
-    assert "snapshots unused" not in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
 
 
@@ -243,156 +240,164 @@ def test_failing_snapshots_diff(stubs, testcases_updated, snapshot):
     testdir.makepyfile(test_file="\n\n".join(testcases_updated.values()))
     result = testdir.runpytest("-vv")
     result_stdout = clean_output(result.stdout.str())
-    start_index = result_stdout.find("==== FAILURES")
-    end_index = result_stdout.find("==== 5 failed")
-    result_stdout = "\n".join(
-        line
-        for line in result_stdout[start_index:end_index].splitlines()
-        if not line.startswith("____") and not line.startswith("====")
-    )
-    assert snapshot == result_stdout
+    start_index = result_stdout.find("\n", result_stdout.find("==== FAILURES"))
+    end_index = result_stdout.find("\n", result_stdout.find("---- snapshot report"))
+    assert result_stdout[start_index:end_index] == snapshot
     assert result.ret == 1
 
 
-def test_updated_snapshots(stubs, testcases_updated):
+def test_updated_snapshots(stubs, testcases_updated, snapshot):
     testdir = stubs[1]
     testdir.makepyfile(test_file="\n\n".join(testcases_updated.values()))
     result = testdir.runpytest("-v", "--snapshot-update")
-    result_stdout = clean_output(result.stdout.str())
-    assert "2 snapshots passed" in result_stdout
-    assert "5 snapshots updated" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
 
 
-def test_unused_snapshots(stubs):
+def test_unused_snapshots(stubs, snapshot):
     _, testdir, tests, _ = stubs
     testdir.makepyfile(test_file="\n\n".join(tests[k] for k in tests if k != "unused"))
     result = testdir.runpytest("-v")
-    result_stdout = clean_output(result.stdout.str())
-    assert "snapshots generated" not in result_stdout
-    assert "6 snapshots passed" in result_stdout
-    assert "1 snapshot unused" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 1
 
 
-def test_unused_snapshots_warning(stubs):
+def test_unused_snapshots_warning(stubs, snapshot):
     _, testdir, tests, _ = stubs
     testdir.makepyfile(test_file="\n\n".join(tests[k] for k in tests if k != "unused"))
     result = testdir.runpytest("-v", "--snapshot-warn-unused")
-    result_stdout = clean_output(result.stdout.str())
-    assert "snapshots generated" not in result_stdout
-    assert "6 snapshots passed" in result_stdout
-    assert "1 snapshot unused" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
 
 
-def test_unused_snapshots_ignored_if_not_targeted_by_testnode_ids(stubs):
-    _, testdir, tests, snapshot_file = stubs
-    testdir.makepyfile(test_file="\n\n".join(tests[k] for k in tests if k != "unused"))
+def test_unused_snapshots_ignored_if_not_targeted_by_testnode_ids(testdir, snapshot):
+    testdir.makefile(".ambr", **{"__snapshots__/other_snapfile": ""})
     testdir.makefile(
-        ".ambr", **{"__snapshots__/other_snapfile": ""},
+        ".py",
+        test_life_uhh_finds_a_way=(
+            """
+            def test_life_always_finds_a_way(snapshot):
+                assert snapshot == snapshot
+
+            def test_clever_girl(snapshot):
+                assert snapshot == snapshot
+            """
+        ),
     )
-    test_content = (
-        "def test_life_always_finds_a_way(snapshot):\n\tassert snapshot == snapshot"
-    )
-    testdir.makefile(
-        ".py", **{"test_life_always_finds_a_way": test_content},
-    )
-    testfile = os.path.join(testdir.tmpdir, "test_life_always_finds_a_way.py")
+    testfile = Path(testdir.tmpdir, "test_life_uhh_finds_a_way.py")
+    testdir.runpytest(str(testfile), "-v", "--snapshot-update")
     result = testdir.runpytest(
         f"{testfile}::test_life_always_finds_a_way", "-v", "--snapshot-update"
     )
-    result_stdout = clean_output(result.stdout.str())
-    assert "1 snapshot generated" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
-    assert os.path.isfile(snapshot_file)
-    assert os.path.isfile("__snapshots__/other_snapfile.ambr")
+    assert Path("__snapshots__/other_snapfile.ambr").exists()
 
 
-def test_unused_snapshots_ignored_if_not_targeted_by_module_testfiles(stubs):
+def test_unused_snapshots_ignored_if_not_targeted_by_module_testfiles(stubs, snapshot):
     _, testdir, tests, _ = stubs
     testdir.makepyfile(test_file="\n\n".join(tests[k] for k in tests if k != "unused"))
-    testdir.makefile(
-        ".ambr", **{"__snapshots__/other_snapfile": ""},
-    )
+    testdir.makefile(".ambr", **{"__snapshots__/other_snapfile": ""})
     result = testdir.runpytest("-v", "--snapshot-update", "test_file.py")
-    result_stdout = clean_output(result.stdout.str())
-    assert "1 unused snapshot deleted" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
-    assert os.path.isfile("__snapshots__/other_snapfile.ambr")
+    assert Path("__snapshots__/other_snapfile.ambr").exists()
 
 
-def test_unused_snapshots_cleaned_up_when_targeting_specific_testfiles(stubs):
-    _, testdir, tests, _ = stubs
+def test_unused_snapshots_cleaned_up_when_targeting_specific_testfiles(stubs, snapshot):
+    _, testdir, _, _ = stubs
     testdir.makepyfile(
         test_file=(
             """
             def test_used(snapshot):
                 assert True
             """
-        )
+        ),
     )
-    testdir.makefile(
-        ".ambr", **{"__snapshots__/other_snapfile": ""},
-    )
+    testdir.makefile(".ambr", **{"__snapshots__/other_snapfile": ""})
     result = testdir.runpytest("-v", "--snapshot-update", "test_file.py")
-    result_stdout = clean_output(result.stdout.str())
-    assert "7 unused snapshots deleted" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
-    assert os.path.isfile("__snapshots__/other_snapfile.ambr")
+    assert Path("__snapshots__/other_snapfile.ambr").exists()
 
 
-def test_removed_snapshots(stubs):
+def test_removed_snapshots(stubs, snapshot):
     _, testdir, tests, filepath = stubs
-    assert os.path.isfile(filepath)
+    assert Path(filepath).exists()
     testdir.makepyfile(test_file="\n\n".join(tests[k] for k in tests if k != "unused"))
     result = testdir.runpytest("-v", "--snapshot-update")
-    result_stdout = clean_output(result.stdout.str())
-    assert "snapshot unused" not in result_stdout
-    assert "1 unused snapshot deleted" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
-    assert os.path.isfile(filepath)
+    assert Path(filepath).exists()
 
 
-def test_removed_snapshot_fossil(stubs):
+def test_removed_snapshot_fossil(stubs, snapshot):
     _, testdir, tests, filepath = stubs
-    assert os.path.isfile(filepath)
+    assert Path(filepath).exists()
     testdir.makepyfile(test_file=tests["inject"])
     result = testdir.runpytest("-v", "--snapshot-update")
-    result_stdout = clean_output(result.stdout.str())
-    assert "snapshots unused" not in result_stdout
-    assert "7 unused snapshots deleted" in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
-    assert not os.path.isfile(filepath)
+    assert not Path(filepath).exists()
 
 
-def test_removed_empty_snapshot_fossil_only(stubs):
+def test_removed_empty_snapshot_fossil_only(stubs, snapshot):
     _, testdir, _, filepath = stubs
     testdir.makefile(".ambr", **{"__snapshots__/test_empty": ""})
-    empty_filepath = os.path.join(testdir.tmpdir, "__snapshots__/test_empty.ambr")
-    assert os.path.isfile(empty_filepath)
+    empty_filepath = Path(testdir.tmpdir, "__snapshots__/test_empty.ambr")
+    assert empty_filepath.exists()
     result = testdir.runpytest("-v", "--snapshot-update")
-    result_stdout = clean_output(result.stdout.str())
-    assert os.path.relpath(filepath) not in result_stdout
-    assert "1 unused snapshot deleted" in result_stdout
-    assert "empty snapshot" in result_stdout
-    assert os.path.relpath(empty_filepath) in result_stdout
+    assert get_result_snapshot_summary(result) == snapshot
     assert result.ret == 0
-    assert os.path.isfile(filepath)
-    assert not os.path.isfile(empty_filepath)
+    assert Path(filepath).exists()
+    assert not Path(empty_filepath).exists()
 
 
-def test_removed_hanging_snapshot_fossil(stubs):
+def test_removed_hanging_snapshot_fossil(stubs, snapshot):
     _, testdir, _, filepath = stubs
     testdir.makefile(".abc", **{"__snapshots__/test_hanging": ""})
-    hanging_filepath = os.path.join(testdir.tmpdir, "__snapshots__/test_hanging.abc")
-    assert os.path.isfile(hanging_filepath)
+    hanging_filepath = Path(testdir.tmpdir, "__snapshots__/test_hanging.abc")
+    assert hanging_filepath.exists()
     result = testdir.runpytest("-v", "--snapshot-update")
     result_stdout = clean_output(result.stdout.str())
-    assert os.path.relpath(filepath) not in result_stdout
+    assert str(Path(filepath).relative_to(Path.cwd())) not in result_stdout
     assert "1 unused snapshot deleted" in result_stdout
     assert "unknown snapshot" in result_stdout
-    assert os.path.relpath(hanging_filepath) in result_stdout
+    assert str(hanging_filepath.relative_to(Path.cwd())) in result_stdout
     assert result.ret == 0
-    assert os.path.isfile(filepath)
-    assert not os.path.isfile(hanging_filepath)
+    assert Path(filepath).exists()
+    assert not Path(hanging_filepath).exists()
+
+
+def test_snapshot_default_extension_option(testdir, snapshot):
+    testdir.makepyfile(
+        test_file=(
+            """
+            def test_default(snapshot):
+                assert b"default extension serializer" == snapshot
+            """
+        ),
+    )
+    result = testdir.runpytest(
+        "-v",
+        "--snapshot-update",
+        "--snapshot-default-extension",
+        "syrupy.extensions.single_file.SingleFileSnapshotExtension",
+    )
+    assert get_result_snapshot_summary(result) == snapshot
+    assert Path(testdir.tmpdir, "__snapshots__/test_file/test_default.raw").exists()
+    assert result.ret == 0
+
+
+def test_snapshot_default_extension_option_failure(testdir, testcases, snapshot):
+    testdir.makepyfile(test_file=testcases["used"])
+    result = testdir.runpytest(
+        "-v",
+        "--snapshot-update",
+        "--snapshot-default-extension",
+        "syrupy.extensions.amber.DoesNotExistExtension",
+    )
+    result_stderr = clean_output(result.stderr.str())
+    assert "error: argument --snapshot-default-extension" in result_stderr
+    assert "Member 'DoesNotExistExtension' not found" in result_stderr
+    assert result.ret